% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_tex.R
\name{optimize_tex}
\alias{optimize_tex}
\title{Optimize TeX File}
\usage{
optimize_tex(file, reduce_power = 0, as_file = FALSE)
}
\arguments{
\item{file}{A character string specifying the file path to be optimized.}

\item{reduce_power}{A parameter to control the reduction of the file size. Default is 0.}

\item{as_file}{A logical value indicating whether to save the plot as a standalone TeX file. Default is FALSE.}
}
\value{
None. The function is called for its side effects.
}
\description{
Optimizes a TeX file by removing unnecessary lines and rounding floats.
}
\examples{
\dontrun{
optimize_tex("path/to/your/file.tex")
}
}
