% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2col.R
\name{rgb2col}
\alias{rgb2col}
\title{Convert RGB Values to the Closest Named R Color}
\usage{
rgb2col(r, g, b)
}
\arguments{
\item{r}{Numeric value for red (0-255)}

\item{g}{Numeric value for green (0-255)}

\item{b}{Numeric value for blue (0-255)}
}
\value{
A string containing the name of the closest R color to the RGB values
}
\description{
This function takes RGB values and finds the closest named R color by
calculating the Euclidean distance between the input RGB values and all
available named colors in R.
}
\examples{
rgb2col(255, 40, 10)  # Returns a color name like "red" or "firebrick"

}
