% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tikz.R
\name{save_tikz}
\alias{save_tikz}
\title{Save Plot as TikZ/TeX file without optimization}
\usage{
save_tikz(g, file, width = 6.9, height = 4.5)
}
\arguments{
\item{g}{A ggplot object to be saved.}

\item{file}{A character string specifying the file path to save the plot.}

\item{width}{The width of the plot in inches. Default is 6.9.}

\item{height}{The height of the plot in inches. Default is 4.5.}
}
\value{
None. The function is called for its side effects.
}
\description{
Saves a ggplot object as a TikZ/TeX file without optimization.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
save_ikz(p, "plot.tex")
}
}
