% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tex.R
\name{save_tex}
\alias{save_tex}
\title{Save ggplot2 Plot as LaTeX (TikZ)}
\usage{
save_tex(g, file, width = 6.9, height = 4.5, reduce_power = 0)
}
\arguments{
\item{g}{A ggplot2 object to be saved.}

\item{file}{The path to the output .tex file.}

\item{width}{The width of the plot in inches. Default is 6.9.}

\item{height}{The height of the plot in inches. Default is 4.5.}

\item{reduce_power}{A parameter to control the reduction of the file size. Default is 5.}
}
\value{
None
}
\description{
Saves a ggplot2 plot as a LaTeX (TikZ) file, optimizing the resulting .tex file by rounding coordinates, removing duplicated elements, and reducing the file size.
}
\examples{
\dontrun{
library(ggplot2)
library(ggplot2latex)
ggplot_object <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()
save_tex(ggplot_object, "path/to/file.tex")
}
}
